Completes RC process for this branch
It is somewhat annoying and uneccessary to have to download all the
extraneous files in the SRFI repo (think the extra tar.gz, all the html
files) for the CHICKEN egg, since it can work without all those extra
files. Instead, I switched to using the files explicitly listed in the
meta file. While there appear to be a set of unrelated commits here, I
needed to perform these commits in order to test Github to see if
henrietta-cache (what downloads / caches the egg once it's tagged) will
still work properly given the specified Github URLs.